requirements.txt
# --- Core Dependencies ---
numpy==1.26.4 \
    --hash=sha256:3908f921d746d8438db02cdd33b5c659a85d38118029c7ae2c90c7936a282f9f
faiss-cpu==1.7.4 \
    --hash=sha256:88995a3962646279f061e8605589146522c7104d49d5641753951f28876c243
kuzu==0.11.3 \
    --hash=sha256:2299a9a396a66279f061e8605589146522c7104d49d5621753951f28876c2d4

# --- Configuration & Serialization ---
pydantic==1.10.15 \
    --hash=sha256:d89481977e8ca686868686868686868686868686868686868686868686868686
PyYAML==6.0.1 \
    --hash=sha256:bfdf3093445cd126868686868686868686868686868686868686868686868686

# --- NLP & Models ---
gliner==0.2.24 \
    --hash=sha256:e89481977e23468686868686868686868686868686868686868686868686868
torch>=2.0.0
transformers>=4.30.0
